leam-base.commons-lang.mid-1140.idx-44426.6.mutant

+++ commons-lang/src/main/java/org/apache/commons/lang3/mutable/MutableLong.java
@@ -122,7 +122,7 @@ public class MutableLong extends Number implements Comparable<MutableLong>, Muta
      * @since 3.5
      */
     public long getAndIncrement() {
-        final long last = value;
+        final long last = (long) value;
         value++;
         return last;
     }
